+++ lib/src/models/errors/azure_errors.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: lines_longer_than_80_chars
+
 import 'package:azure_app_config/src/models/errors/error_response.dart';
 
 /// When an KeyValue cannot be parsed to a FeatureFlag.