+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+## 0.13.0
+
+- Complete Azure's Default 'Targeting' filter by implementing the 'groups' and 'users' feature. 
+
 ## 0.12.0
+
 - Add registering of ComplexTypes for automatic (de)serialization of models.
 
 ## 0.11.0